Rename & re-organize FPTIBatchData structs
